Fix data gaps on Bar Chart
This chunk of code is necessary to show data gaps in charts that have an expected cadence of data.  For line charts this was working fine, but for Bar Charts, putting the null data point at a millisecond time interval was screwing up Rechart's math on how wide it could make each bar.