MostRecentNotification: Add identifier regex filter (#401)
* MostRecentNotification: Add identifier regex filter

This adds a new property to the MostRecentNotification component: A regex string that will filter the notifications. It works like the NotificationList component.

One major change I had to make here was how we fetched notifications. If the notification identifier regex is supplied, we have to query until something matches or we run out of results.

Another change I made here is brough the storybook for this component up to date. I followed the example set by NotificationList component.

* remove a console.log

* fix preview state behavior

* simplify iteration over notifications

* simplify iteration over notifications

* more tweaks to request logic

* Fix issue with preview state

* validate regex